financial_agent/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
aws-requests-auth==0.4.3
blinker==1.9.0
boto3==1.38.42
botocore==1.38.42
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dill==0.4.0
docstring_parser==0.16
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
mpmath==1.3.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
pillow==11.2.1
prompt_toolkit==3.0.51
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
rich==14.0.0
s3transfer==0.13.0
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.47.1
strands-agents==0.1.8
strands-agents-tools==0.1.6
sympy==1.14.0
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.23.0
